Split NoSudo between Debian and RedHat
